refactored code 로컬 테스트 완료
