+++ src/etl/crawling/book_url_getter.py
@@ -1,7 +1,7 @@
 import os, json
 from bs4 import BeautifulSoup
 
-from etl.utils.config import get_workdir
+from src.etl.utils.config import get_workdir
 
 
 class BookURLGetter:
@@ -12,7 +12,13 @@ class BookURLGetter:
         self.book_category_url = {}
         self.book_page_url = []
 
-    def get_category(self) -> list:
+    def get_book_category_url(self):
+        return self.book_category_url
+    
+    def get_book_page_url(self):
+        return self.book_page_url
+
+    def get_category_scrapper(self) -> list:
         """
         static json file로 카페고리 별 url 정보 저장되어있는 데이터 리턴
         """
@@ -20,22 +26,21 @@ class BookURLGetter:
         with open(os.path.join(self.workdir, book_category_filename)) as f:
             url = json.load(f)
         self.book_category_url = url[self.category]
-
-    def get_book_page_urls(self) -> list:
+        
+    def get_book_page_urls_scrapper(self) -> list:
         """
         각 책의 상세 웹페이지 주소 추출, 리스트로 리턴
         """
-        self.get_category()
-        print(self.book_category_url)
+        self.get_category_scrapper()
         for category_name, category_url in self.book_category_url.items():
             for page in range(1, 4):
                 search_url = category_url+f'&tab=top100&list_type=list&sort_type=publishday&page={page}'
                 self.driver.get(search_url)
-                book_page_urls = self.get_book_page_urls(BeautifulSoup(self.driver.page_source, 'html.parser'))
+                book_page_urls = self.get_book_page_url_scrapper(BeautifulSoup(self.driver.page_source, 'html.parser'))
                 self.book_page_url.append((category_name, page, book_page_urls))
 
     @staticmethod
-    def get_book_page_urls(bsObject):
+    def get_book_page_url_scrapper(bsObject):
         """
         책의 상세 웹페이지 주소를 추출, 리턴
         """
@@ -48,7 +53,3 @@ class BookURLGetter:
             book_page_urls.append(link)
 
         return book_page_urls
-
-if __name__ == "__main__":
-    test = BookURLGetter("소설")
-    test.get_book_page_urls()
